Make -base compile almost without warnings again.  Fixes:
- Using the correct format specifier
- Remove substringFromRange: (GNUstep extension) from NSString.h and use the non-deprecated version in -base.
- Delete definitions of methods that are the same in both a class and a category on that class (NSArray and NSObject).

Two files still have warnings:

- The GNUTLS stuff is using a load of deprecated APIs.
- xmlparse.m replaces a method using a category.  This method seems to always return nil.  Not sure what it's meant to do, or why it isn't a delegate of some kind.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@33756 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSPrivate.h
@@ -75,7 +75,7 @@ typedef struct objc_category* Category;
   unsigned	_count;
   unsigned	_capacity;
   int		_grow_factor;
-  int		_version;
+  unsigned long		_version;
 }
 @end
